restructured module hierarchy + renamed data-visibility to data-visible
