+++ technical-explanation.md
@@ -2,22 +2,22 @@
 
 In contract.html, I made two custom HTML attributes, to avoid having to hardcode what data goes in what field and when certain elements are visible. These are:
 
-* `data-visibility`: Determines whether an element should be visible depending on data from form.
+* `data-visible`: Determines whether an element should be visible depending on form data.
 * `data-fill`: Automatically fills content of field with data from form.
 
 Both have their own specific syntax.
 
-### Syntax of the `data-visibility` attribute
+### Syntax of the `data-visible` attribute
 
-The `data-visibility` attribute contains a boolean algebra to help handle complex visibility requirements. The syntax of this algebra is as follows:
+The `data-visible` attribute contains a boolean algebra to help handle complex visibility requirements. The syntax of this algebra is as follows:
 
 * ` `
-  * Logical conjugation (AND).
-  * Whitespaces between operators and their operands are ignored.
+  * Logical conjugation operator (AND).
+  * Whitespaces between operators and their operands are ignored, as are whitespaces immediately on the inside of parentheses.
 * `|`
-  * Logical disjunction (OR).
+  * Logical disjunction operator (OR).
 * `!`
-  * Logical negation (NOT).
+  * Logical negation operator (NOT).
 * `(` and `)`
   * Used to associate subexpressions.
 
@@ -33,18 +33,18 @@ All other symbols are concatenated into atomic expressions. Hence, atomic expres
 The `data-fill` tag has following syntax:
 
 * `|`
-  * Pipe operator. Pipes output of the preceding expression in into the expression after it.
-  * The expression preceding the first pipe is a comma-separated list of keys of the data object. The values of these keys is retrieved and passed into the next operator.
+  * Pipe operator. Pipes output of the preceding expression into the expression after it.
+  * The expression preceding the first pipe is a comma-separated list of keys of the data object. The values of these keys are retrieved and passed into the next expression.
   * The expression following a pipe operator begins with the name of a formatter. The formatter will be applied to the output of the preceding expression.
 * `(` and `)`
   * Some formatters may accept additional parameters. These should be placed between parentheses after the name of the formatter.
   * Parameters are always read as strings. Note that these cannot contain any of these syntactic tokens. (Serious limitation of the language, but I can't be arsed to improve it.)
 * `,`
   * Separates different keys of the data object in the first subexpression.
-  * Separates parameters of the formatter.
+  * Separates parameters of a formatter.
 * ` `
-  * Trimmed away when found around operators. (Also potentially a limitation.)
- 
+  * Trimmed away when found around operators. (Also potentially a limitation.) 
+
 If the final result is a list, the contained elements will be joined with spaces.
 
 Available formatters include:
@@ -58,6 +58,7 @@ Available formatters include:
 * `address`
   * Takes the values streetName, houseNumber, boxNumber, postalCode, municipality, and country.
   * Formats these into an address, according to Belgian conventions.
+  * If both the streetname and municipality are empty, returns empty string.
 * `telephone`
   * Formats a phone number according to Belgian conventions.
 * `euro`
@@ -70,3 +71,11 @@ Available formatters include:
   * Joins values with a comma and a space.
 * `capitalize`
   * Capitalizes first letter of value.
+
+_Example:_ `phoneNumber|telephone`: Retrieves the `phoneNumber` property of the form data object, and formats it as a telephone number.
+
+_Example:_ `firstName, lastName`: Retrieves the `firstName` and `lastName` properties of the form data object, and since there is no formatter specified, joins these with spaces.
+
+_Example:_ `streetName,houseNumber,boxNumber,postalCode,municipality,country|address`: Retrieves various properties from the form data object, and parses these into an address.
+
+_Example:_ `oldIncludesCharger, oldIncludesMouse, oldIncludesSmartCardReader|mask(lader, muis, eID-lezer)|joincomma|capitalize`: Retrieves properties from the form data object, masks the list ['lader', 'muis', 'eID-lezer'] according to their values, joins the remaining values from this list with `, `, and capitalizes the first character.
